moving to a slow test
